- fixes comparison of struct against null (no no no) - fixes IRCBridgeModule's XmlRpc method really paying attention to   region parameter - cleans up indentation in IRCBridge code - fixes ConciergeModule exception on client logout

+++ OpenSim/Region/Environment/Modules/Avatar/Chat/IRCBridgeModule.cs
@@ -241,7 +241,7 @@ namespace OpenSim.Region.Environment.Modules.Avatar.Chat
                     if (rs.Region == region)
                     {
                         responseData["server"]    = rs.cs.Server;
-                        responseData["port"]      = rs.cs.Port;
+                        responseData["port"]      = (int)rs.cs.Port;
                         responseData["user"]      = rs.cs.User;
                         responseData["channel"]   = rs.cs.IrcChannel;
                         responseData["enabled"]   = rs.cs.irc.Enabled;
